simplify apps/cli/src/dashboard/
