+++ apps/cli/src/commands/watch-dashboard.tsx
@@ -5,117 +5,14 @@
  * a modern terminal-based dashboard interface built with React/Ink.
  */
 
-import React, { useState, useEffect } from 'react';
-import { render, useApp, useInput } from 'ink';
-import { ServiceType, isServiceType } from '../lib/types.js';
-import { DashboardDataSource, DashboardData } from '../dashboard/dashboard-data.js';
-import { UnifiedDashboard, LogsOnlyDashboard, MetricsOnlyDashboard } from '../dashboard/dashboard-layouts.js';
-
-// Dashboard mode types
-type DashboardMode = 'unified' | 'logs' | 'metrics';
-
-// Main Dashboard App Component
-const DashboardApp: React.FC<{ 
-  mode: DashboardMode; 
-  service?: ServiceType; 
-  refreshInterval?: number;
-  environment: string;
-  data?: DashboardData;  // Accept data as a prop
-}> = ({ mode, service, refreshInterval = 30, environment, data: propData }) => {
-  // Use prop data if provided, otherwise maintain own state
-  const [data, setData] = useState<DashboardData>(propData || {
-    services: [],
-    logs: [],
-    metrics: [],
-    lastUpdate: new Date(),
-    isRefreshing: false
-  });
-  const [dataSource] = useState(() => propData ? null : new DashboardDataSource(environment));
-  const { exit } = useApp();
-
-  // Global keyboard shortcuts
-  useInput((input, key) => {
-    if (input === 'q' || (key.ctrl && input === 'c')) {
-      exit();
-    } else if (input === 'r') {
-      refreshData();
-    }
-  });
-
-  // Data refreshing
-  const refreshData = async () => {
-    // Only refresh if we have our own dataSource (not using prop data)
-    if (!dataSource) return;
-    
-    setData(prev => ({ ...prev, isRefreshing: true }));
-    try {
-      const newData = await dataSource.getDashboardData();
-      setData(newData);
-    } catch (error) {
-      console.error('Failed to refresh dashboard data:', error);
-    } finally {
-      setData(prev => ({ ...prev, isRefreshing: false }));
-    }
-  };
-
-  // Update data when prop changes
-  useEffect(() => {
-    if (propData) {
-      setData(propData);
-    }
-  }, [propData]);
-
-  // Initial load and periodic refresh (only if not using prop data)
-  useEffect(() => {
-    if (!propData) {
-      refreshData();
-      const interval = setInterval(refreshData, refreshInterval * 1000);
-      return () => clearInterval(interval);
-    }
-  }, [refreshInterval, propData]);
-
-  // Filter data by service if specified
-  const filteredLogs = service && service !== 'both' 
-    ? data.logs.filter(log => log.service === service)
-    : data.logs;
-
-  const filteredServices = service && service !== 'both'
-    ? data.services.filter(s => s.name.toLowerCase().includes(service))
-    : data.services;
-
-  // Render appropriate dashboard layout
-  switch (mode) {
-    case 'unified':
-      return <UnifiedDashboard data={data} refreshInterval={refreshInterval} />;
-      
-    case 'logs':
-      return (
-        <LogsOnlyDashboard 
-          logs={filteredLogs}
-          services={filteredServices}
-          title={service ? `${service.charAt(0).toUpperCase() + service.slice(1)} Logs` : 'All Service Logs'}
-          {...(service && service !== 'both' && { serviceFilter: [service] })}
-        />
-      );
-      
-    case 'metrics':
-      return (
-        <MetricsOnlyDashboard 
-          metrics={data.metrics}
-          services={data.services}
-          title="Performance Metrics"
-        />
-      );
-      
-    default:
-      return <UnifiedDashboard data={data} refreshInterval={refreshInterval} />;
-  }
-};
+import React from 'react';
+import { render } from 'ink';
+import { DashboardApp, DashboardMode } from '../dashboard/dashboard-layouts.js';
 
 import { getAvailableEnvironments } from '../platforms/platform-resolver.js';
 
 // Argument parsing with environment support
-function parseArgs(): { environment: string; mode: DashboardMode; service?: ServiceType } {
+function parseArgs(): { environment: string; mode: DashboardMode } {
   const args = process.argv.slice(2);
   
   if (args.length === 0 || args[0] === 'help' || args[0] === '--help' || args[0] === '-h') {
@@ -136,61 +33,39 @@ function parseArgs(): { environment: string; mode: DashboardMode; service?: Serv
   }
   
   let mode: DashboardMode = 'unified';
-  let service: ServiceType | undefined;
 
   // Parse mode from remaining args
   const remainingArgs = args.slice(1);
   if (remainingArgs[0] === 'logs') {
     mode = 'logs';
-    if (remainingArgs[1] && isServiceType(remainingArgs[1])) {
-      service = remainingArgs[1];
-    }
   } else if (remainingArgs[0] === 'metrics') {
     mode = 'metrics';
-    if (remainingArgs[1] && isServiceType(remainingArgs[1])) {
-      service = remainingArgs[1];
-    }
-  } else if (remainingArgs[0] && isServiceType(remainingArgs[0])) {
-    // Service specified first
-    service = remainingArgs[0];
-    if (remainingArgs[1] === 'logs') {
-      mode = 'logs';
-    } else if (remainingArgs[1] === 'metrics') {
-      mode = 'metrics';
-    }
   } else if (remainingArgs.length > 0 && remainingArgs[0] !== 'unified') {
     console.error(`Unknown argument: ${remainingArgs[0]}`);
     showHelp();
   }
 
-  return { environment, mode, ...(service && { service }) };
+  return { environment, mode };
 }
 
 function showHelp(): never {
   console.log('Semiont Watch - Interactive System Dashboard\n');
-  console.log('Usage: semiont watch <environment> [mode] [service]\n');
+  console.log('Usage: semiont watch <environment> [mode]\n');
   
   console.log('Arguments:');
   console.log(`  <environment>  Environment to watch (${getAvailableEnvironments().join(', ')})`);
-  console.log('  [mode]         Dashboard mode (default: unified)');
-  console.log('  [service]      Service filter (default: all)\n');
+  console.log('  [mode]         Dashboard mode (default: unified)\n');
   
   console.log('Modes:');
   console.log('  unified   Unified dashboard with services, logs, and metrics (default)');
   console.log('  logs      Focus on log streaming');
   console.log('  metrics   Focus on performance metrics\n');
   
-  console.log('Services:');
-  console.log('  frontend  Filter to frontend service only');
-  console.log('  backend   Filter to backend service only');
-  console.log('  (none)    Show all services\n');
   
   console.log('Examples:');
   console.log('  semiont watch production               # Unified dashboard');
   console.log('  semiont watch staging logs             # All service logs');
-  console.log('  semiont watch development logs frontend # Frontend logs only');
-  console.log('  semiont watch production metrics        # Performance dashboard');
-  console.log('  semiont watch staging frontend          # Frontend unified view\n');
+  console.log('  semiont watch production metrics        # Performance dashboard\n');
   
   console.log('Controls:');
   console.log('  q         Quit');
@@ -205,7 +80,7 @@ function showHelp(): never {
 // Main execution
 async function main() {
   try {
-    const { environment, mode, service } = parseArgs();
+    const { environment, mode } = parseArgs();
     
     // Get refresh interval from environment variable if set
     const refreshInterval = process.env.SEMIONT_REFRESH_INTERVAL 
@@ -216,7 +91,6 @@ async function main() {
     render(
       <DashboardApp 
         mode={mode} 
-        {...(service && { service })}
         refreshInterval={refreshInterval}
         environment={environment}
       />
@@ -236,3 +110,4 @@ if (import.meta.url === `file://${process.argv[1]}`) {
 
 // Export for use in watch.ts
 export default DashboardApp;
+export { DashboardMode };
